kmalloc, need to add 64bit for printk

+++ src/Rpi lib/testing_lib/helper_macros.h
@@ -4,7 +4,7 @@
 
 // powers of 2
 #define is_aligned(x, a)        (((x) & ((typeof(x))(a) - 1)) == 0)
-#define is_aligned_ptr(x, a)        is_aligned((uint64_t)x,a)
+#define is_aligned_ptr(x, a)        is_aligned((uintptr_t)x,a)
 
 #define pi_roundup(x,n) (((x)+((n)-1))&(~((n)-1)))
